work on indexing downloaded aligned bibles
